pyproject.toml
[project]
name = "airmonitoring"
version = "0.1.0"
description = "AirGradient data collection and analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "marimo[lsp,recommended]>=0.15.0",
    "python-dotenv",
    "duckdb<1.4",
    "polars",
    "pyarrow",
    "requests",
    "typer",
    "vegafusion[embed]",
    "vl-convert-python",
    "ipykernel",
    "numpy>=2.3.4",
]

[dependency-groups]
ci = [
    "requests",
    "python-dotenv",
    "duckdb<1.4",
    "polars",
    "pyarrow",
    "typer",
]
